Move login bottom sheet under Main
