+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/screens/account/generating/GeneratingScreen.kt
@@ -19,8 +19,8 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import kotlinx.coroutines.delay
 import net.nymtech.nymvpn.R
 import net.nymtech.nymvpn.ui.Route
-import net.nymtech.nymvpn.ui.screens.auth.AuthRoute
-import net.nymtech.nymvpn.ui.screens.auth.routeName
+import net.nymtech.nymvpn.ui.AuthRoute
+import net.nymtech.nymvpn.ui.routeName
 import net.nymtech.nymvpn.ui.common.animations.PulsingDotsWave
 import net.nymtech.nymvpn.ui.common.navigation.LocalNavController
 import net.nymtech.nymvpn.ui.common.snackbar.AlertController
@@ -29,6 +29,7 @@ import net.nymtech.nymvpn.ui.common.snackbar.AlertType
 import net.nymtech.nymvpn.ui.theme.*
 import net.nymtech.nymvpn.util.extensions.navigateAndForget
 import net.nymtech.nymvpn.util.extensions.replaceCurrentWith
+import kotlin.time.Duration.Companion.milliseconds
 
 @Composable
 fun GeneratingScreen(viewModel: GeneratingViewModel = hiltViewModel()) {
@@ -80,10 +81,10 @@ fun GeneratingContent(mode: GeneratingMode, onAnimationEnd: () -> Unit) {
 	LaunchedEffect(isDeepLink) {
 		if (!isDeepLink) {
 			repeat(2) {
-				delay(3000)
+				delay(3000.milliseconds)
 				step++
 			}
-			delay(3000)
+			delay(3000.milliseconds)
 			onAnimationEnd()
 		}
 	}
