+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/screens/main/MainModals.kt
@@ -3,25 +3,21 @@ package net.nymtech.nymvpn.ui.screens.main
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
-import androidx.lifecycle.lifecycleScope
-import kotlinx.coroutines.launch
 import net.nymtech.nymvpn.R
 import net.nymtech.nymvpn.ui.AppUiState
-import net.nymtech.nymvpn.ui.MainActivity
 import net.nymtech.nymvpn.ui.common.snackbar.AlertController
 import net.nymtech.nymvpn.ui.common.snackbar.AlertMessage
 import net.nymtech.nymvpn.ui.common.snackbar.AlertType
 import net.nymtech.nymvpn.ui.screens.account.info.AutologinState
 import net.nymtech.nymvpn.ui.screens.account.info.modal.AutologinLoadingDialog
 import net.nymtech.nymvpn.ui.screens.account.info.modal.PinCodeDialog
-import net.nymtech.nymvpn.ui.screens.auth.AuthBottomSheet
-import net.nymtech.nymvpn.ui.screens.auth.MainBottomSheetContent
+import net.nymtech.nymvpn.ui.screens.main.bottomsheet.MainBottomSheet
+import net.nymtech.nymvpn.ui.screens.main.bottomsheet.MainBottomSheetContent
 import net.nymtech.nymvpn.ui.screens.main.modal.BatteryModal
 import net.nymtech.nymvpn.ui.screens.main.modal.CompatibilityModal
 import net.nymtech.nymvpn.ui.screens.main.modal.NetworkStatsModal
 import net.nymtech.nymvpn.ui.screens.main.modal.NodeFamiliesModal
 import net.nymtech.nymvpn.ui.screens.main.modal.ShowInfoModal
-import net.nymtech.nymvpn.util.extensions.savePasswordToManager
 
 @Composable
 fun MainModals(
@@ -52,7 +48,6 @@ fun MainModals(
 	appUiState: AppUiState,
 ) {
 	val context = LocalContext.current
-	val activity = context as? MainActivity
 
 	when (val autologin = autologinState) {
 		is AutologinState.Loading -> AutologinLoadingDialog(onCancel = onCancelAutologin)
@@ -103,16 +98,11 @@ fun MainModals(
 		onDismiss = onDismissNodeFamilies,
 	)
 
-	AuthBottomSheet(
+	MainBottomSheet(
 		content = bottomSheetContent,
 		onDismissRequest = onDismissBottomSheet,
 		onAuthSuccess = onAuthSuccess,
 		onLoginProcessingStart = onLoginProcessingStart,
-		onSaveToPasswordManager = { password ->
-			activity?.lifecycleScope?.launch {
-				savePasswordToManager(context = context, password = password)
-			}
-		},
 		authSheetMinHeightPx = authSheetMinHeightPx,
 		onAuthSheetHeightChange = onAuthSheetHeightChange,
 		appUiState = appUiState,
